requirements.txt
absl-py==1.4.0
accelerate==0.21.0
add-parent-path==0.1.11
aiofiles==23.1.0
aiogoogletrans==3.3.3
aiohttp==3.8.5
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audio-recorder-streamlit==0.0.8
av==10.0.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
bokeh==3.2.0
breadability==0.1.20
cachetools==5.3.1
certifi==2024.2.2
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.4
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.1.0
cryptography==41.0.2
ctransformers==0.2.27
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deta==1.1.0
diffusers==0.19.3
diskcache==5.6.3
distlib==0.3.5
dnspython==2.3.0
docopt==0.6.2
effdet==0.4.1
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==0.9.1
faiss-cpu==1.8.0
fastapi==0.100.1
fastjsonschema==2.16.1
ffmpy==0.3.1
filelock==3.7.1
filetype==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gistory==0.44
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-language==2.10.1
google-cloud-speech==2.21.0
google-crc32c==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
googletrans==3.1.0a0
gradio==3.39.0
gradio_client==0.3.0
greenlet==3.0.3
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.9.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.21.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
importlib-metadata==6.7.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
JPype1==1.4.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kanilog==0.3.3
keras==2.13.1
kiwisolver==1.4.4
konlpy==0.6.0
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.13
layoutparser==0.3.4
libclang==16.0.6
linkify-it-py==2.0.2
llama_cpp_python==0.2.54
loguru==0.7.2
logzero==1.7.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.21.0
matplotlib==3.7.2
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mecab-python @ file:///C:/my_develop2/AI_TBM/myenv/Lib/site-packages/mecab_python-0.996_ko_0.9.2_msvc-cp310-cp310-win_amd64.whl#sha256=5f323652ee6a24beb9319ce41009874cc8c2c62c9cd045c5421291ddecea00f4
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.8.1
notebook==6.4.12
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pickleshare==0.7.5
pikepdf==8.13.0
pillow==10.2.0
pillow_heif==0.15.0
platformdirs==2.5.2
pluggy==1.2.0
portalocker==2.8.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.1
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycocotools==2.0.7
pycountry==22.3.5
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.0.0
Pygments==2.15.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypdf==4.0.2
pypdfium2==4.27.0
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.19.3
pyshorteners==1.0.1
pytesseract==0.3.10
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyunsplash==1.0.0rc2
pywin32==304
pywinpty==2.0.7
PyYAML==6.0
pyzmq==23.2.0
rapidfuzz==3.6.1
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.8.8
rsa==4.9
safetensors==0.4.2
scikit-learn==1.1.2
scipy==1.11.1
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.5.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.27
stack-data==0.3.0
starlette==0.27.0
stdlogging==0.16
streamlit==1.31.1
streamlit-audiorecorder==0.0.2
streamlit-bokeh-events==0.1.2
streamlit-lottie==0.0.5
streamlit-webrtc==0.45.1
sumy==0.11.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.15.0
threadpoolctl==3.1.0
timm==0.9.16
tinycss2==1.1.1
tmdbv3api==1.7.7
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.1
torchvision==0.17.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.3.0
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==4.2
uc-micro-py==1.0.2
unstructured==0.12.5
unstructured-client==0.21.0
unstructured-inference==0.7.24
urllib3==2.2.1
uvicorn==0.23.1
validators==0.20.0
virtualenv==20.16.2
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==3.6.1
win32-setctime==1.1.0
wrapt==1.15.0
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.15.0
